feat: sync adding a chapter server/client
