+++ common/src/main/java/fr/loudo/narrativecraft/platform/services/IPacketSender.java
@@ -28,5 +28,7 @@ import net.minecraft.server.level.ServerPlayer;
 
 public interface IPacketSender {
 
-    void sendToPlayer(ServerPlayer player, CustomPacketPayload payload);
+    void sendToPlayer(ServerPlayer player, CustomPacketPayload packet);
+
+    void sendToServer(CustomPacketPayload packet);
 }
