+++ neoforge/src/main/java/fr/loudo/narrativecraft/network/server/ServerPacketHandlerNeoForge.java
@@ -21,21 +21,17 @@
  * SOFTWARE.
  */
 
-package fr.loudo.narrativecraft.events;
+package fr.loudo.narrativecraft.network.server;
 
-import fr.loudo.narrativecraft.NarrativeCraftMod;
-import fr.loudo.narrativecraft.network.common.CommonPackerHandlerNeoForge;
 import fr.loudo.narrativecraft.network.data.BiChapterDataPacket;
-import net.neoforged.api.distmarker.Dist;
-import net.neoforged.bus.api.SubscribeEvent;
-import net.neoforged.fml.common.EventBusSubscriber;
-import net.neoforged.neoforge.client.network.event.RegisterClientPayloadHandlersEvent;
+import fr.loudo.narrativecraft.network.handlers.ServerPacketHandler;
+import net.minecraft.server.level.ServerPlayer;
+import net.neoforged.neoforge.network.handling.IPayloadContext;
 
-@EventBusSubscriber(modid = NarrativeCraftMod.MOD_ID, value = Dist.CLIENT)
-public class PacketClientRegisterEvent {
-
-    @SubscribeEvent
-    private static void onPackerRegister(RegisterClientPayloadHandlersEvent event) {
-        event.register(BiChapterDataPacket.TYPE, CommonPackerHandlerNeoForge::chapterData);
+public class ServerPacketHandlerNeoForge {
+    public static void chapterData(BiChapterDataPacket packet, IPayloadContext context) {
+        context.enqueueWork(() -> {
+            ServerPacketHandler.chapterData(packet, (ServerPlayer) context.player());
+        });
     }
 }
