+++ neoforge/src/main/java/fr/loudo/narrativecraft/platform/NeoForgePacketSender.java
@@ -26,11 +26,17 @@ package fr.loudo.narrativecraft.platform;
 import fr.loudo.narrativecraft.platform.services.IPacketSender;
 import net.minecraft.network.protocol.common.custom.CustomPacketPayload;
 import net.minecraft.server.level.ServerPlayer;
+import net.neoforged.neoforge.client.network.ClientPacketDistributor;
 import net.neoforged.neoforge.network.PacketDistributor;
 
 public class NeoForgePacketSender implements IPacketSender {
     @Override
-    public void sendToPlayer(ServerPlayer player, CustomPacketPayload payload) {
-        PacketDistributor.sendToPlayer(player, payload);
+    public void sendToPlayer(ServerPlayer player, CustomPacketPayload packet) {
+        PacketDistributor.sendToPlayer(player, packet);
+    }
+
+    @Override
+    public void sendToServer(CustomPacketPayload packet) {
+        ClientPacketDistributor.sendToServer(packet);
     }
 }
